requirements.txt
gunicorn==19.3.0
Flask==0.12
protobuf==3.0.0
psutil==3.4.2
ptyprocess==0.5
py==1.4.31
pyflakes==1.1.0
Pygments==2.1
pylint==1.5.2
pyparsing==2.0.3
python-dateutil==2.4.2
python-geoip==1.2
python-geoip-geolite2==2015.303
requests==2.11.1
Werkzeug==0.11.15
wrapt==1.8.0
xlrd==0.9.4
XlsxWriter==0.9.4
xlwt==0.7.5
xvfbwrapper==0.2.8
